projects/02-inbox-pilot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.6.1
google-auth==2.25.2
google-api-python-client==2.111.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
requests==2.31.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.23.2
